+++ msmbuilder/decomposition/__init__.py
@@ -3,3 +3,4 @@ from __future__ import absolute_import
 from .pca import PCA, SparsePCA, MiniBatchSparsePCA
 from .sparsetica import SparseTICA
 from .tica import tICA
+from .ktica import KernelTICA