+++ msmbuilder/feature_extraction/__init__.py
@@ -1,0 +1,4 @@
+from __future__ import absolute_import
+
+from ..featurizer import *
+from ..featurizer import subset