+++ msmbuilder/feature_selection/base.py
@@ -1,0 +1,12 @@
+# Author: Carlos Xavier Hernandez <cxh@stanford.edu>
+# Contributors:
+# Copyright (c) 2016, Stanford University and the Authors
+# All rights reserved.
+
+from __future__ import absolute_import
+
+from ..decomposition.base import MultiSequenceDecompositionMixin
+
+
+class MultiSequenceFeatureSelectionMixin(MultiSequenceDecompositionMixin):
+    __doc__ = MultiSequenceDecompositionMixin.__doc__