+++ msmbuilder/tests/test_lumping.py
@@ -6,12 +6,13 @@ from sklearn.pipeline import Pipeline
 from msmbuilder.lumping import PCCA, PCCAPlus
 from msmbuilder.msm import MarkovStateModel
 
+random = np.random.RandomState(9)
+
 
 def _metastable_system():
     # Make a simple dataset with four states, where there are 2 obvious
     # macrostate. States 0 and 1 interconvert, states 2 and 3 interconvert.
     n = 250
-    random = np.random.RandomState(1)
     assignments = np.concatenate([
         0 + random.randint(2, size=n),  # states 0, 1 interconverting
         2 + random.randint(2, size=n),  # states 2, 3 intercoverting
@@ -30,11 +31,11 @@ def _metastable_system():
     return assignments, macro_assignments
 
 
-def test_pcca_1():
+def test_pcca():
     assignments, ref_macrostate_assignments = _metastable_system()
     pipeline = Pipeline([
         ('msm', MarkovStateModel()),
-        ('pcca+', PCCA(2))
+        ('pcca', PCCA(2))
     ])
     macro_assignments = pipeline.fit_transform(assignments)[0]
 
@@ -46,7 +47,7 @@ def test_pcca_1():
             np.all(macro_assignments == opposite))
 
 
-def test_pcca_plus_1():
+def test_pcca_plus():
     assignments, ref_macrostate_assignments = _metastable_system()
     pipeline = Pipeline([
         ('msm', MarkovStateModel()),
@@ -74,28 +75,31 @@ def test_from_msm():
 
 def test_ntimescales_1():
     # see issue #603
-    trajs = [np.random.randint(0, 30, size=500) for _ in range(15)]
+    trajs = [random.randint(0, 100, size=500) for _ in range(15)]
 
     pccap = PCCAPlus(n_macrostates=11).fit(trajs)
     lumped_trajs = pccap.transform(trajs)
-    assert len(np.unique(lumped_trajs)) == 11
+    observed_macros = len(np.unique(lumped_trajs))
+    assert observed_macros == 11, observed_macros
 
 
 def test_ntimescales_2():
     # see issue #603
-    trajs = [np.random.randint(0, 30, size=500) for _ in range(15)]
+    trajs = [random.randint(0, 100, size=500) for _ in range(15)]
     msm = MarkovStateModel().fit(trajs)
 
     pccap = PCCAPlus.from_msm(msm, 11)
     lumped_trajs = pccap.transform(trajs)
-    assert len(np.unique(lumped_trajs)) == 11
+    observed_macros = len(np.unique(lumped_trajs))
+    assert observed_macros == 11, observed_macros
 
 
 def test_ntimescales_3():
     # see issue #603
-    trajs = [np.random.randint(0, 30, size=500) for _ in range(15)]
+    trajs = [random.randint(0, 100, size=500) for _ in range(15)]
     msm = MarkovStateModel(n_timescales=10).fit(trajs)
 
     pccap = PCCAPlus.from_msm(msm, 11)
     lumped_trajs = pccap.transform(trajs)
-    assert len(np.unique(lumped_trajs)) == 11
+    observed_macros = len(np.unique(lumped_trajs))
+    assert observed_macros == 11, observed_macros