+++ docs/changelog.rst
@@ -11,9 +11,22 @@ This is the current development version of MSMBuilder
 API Changes
 ~~~~~~~~~~~
 
+- ``msmbuilder.featurizer.FeatureUnion`` is now deprecated. Please use
+  ``msmbuilder.feature_selection.FeatureSelector`` instead (#799).
+
+- ``msmbuilder.feature_extraction`` has been added to conform to the
+  ``scikit-learn`` API. This is essentially an alias of
+  ``msmbuilder.featurizer`` (#799).
+
+
 
 New Features
 ~~~~~~~~~~~~
+ - ``KernelTICA``, ``Nystroem``, and ``LandmarkNystroem`` are available in the
+   ``decomposition`` module (#807).
+
+ - ``FeatureSelector`` and ``VarianceThreshold`` are available in the
+   ``feature_selection`` module (#799)
 
  - ``SparsePCA`` and ``MiniBatchSparsePCA`` are available in the
    ``decomposition`` module (#791).